tb-detector/requirements.txt
flask>=3.0
flask-cors>=4.0
numpy>=1.24
pillow>=10.0
tensorflow>=2.15
gunicorn>=21.0
